Split NDJSON line writing
